backend/requirements.txt
fastapi==0.115.11
uvicorn[standard]==0.34.0
pydantic==2.10.6
pydantic-settings==2.8.1
sqlalchemy==2.0.38
alembic==1.14.1
psycopg[binary]==3.2.6
python-jose==3.4.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.20

